Add iCloud Private Relay domains to general whitelist
e.g. mask.apple-dns.net, mask.icloud.com, mask-api.icloud.com, etc...

Ensuring these go through and are not accidentally blocked by scripts trying to block Apple tracking domains as they are for a privacy preserving service.

Of course as this service is essentially a proxy that redirects (some of) your traffic through Apple, this means placing trust in Apple and Akamai (the CDN they use) to not log that traffic. If you don't trust Apple with your traffic, rather then trying to play whack-a-mole with their domains, it's easier to just disable Private Relay in iCloud settings for your iOS and macOS devices instead.

I assume therefore that anyone who has this feature enabled has made a proactive choice to use it (especially as it requires a paid iCloud tier) and doesn't want the domains blocked on their network.